+++ src/bittytax/bt_types.py
@@ -37,6 +37,7 @@ class TrType(Enum):
     FEE_REBATE = "Fee-Rebate"
     LOAN = "Loan"
     MARGIN_GAIN = "Margin-Gain"
+    MARGIN_FEE_REBATE = "Margin-Fee-Rebate"
     WITHDRAWAL = "Withdrawal"
     SPEND = "Spend"
     GIFT_SENT = "Gift-Sent"
@@ -95,6 +96,7 @@ BUY_TYPES = (
     TrType.FEE_REBATE,
     TrType.LOAN,
     TrType.MARGIN_GAIN,
+    TrType.MARGIN_FEE_REBATE,
     TrType.TRADE,
 )